requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

django-allauth
django-dotenv
django-rest-auth
django==3.0.4
djangorestframework
djangorestframework-jwt
gunicorn==20.0.4
pillow
pipenv-to-requirements
psycopg2
pygments
